Separate NetworkSpy
