+++ Sources/desafioTests/Scenes/Advertisements/APIAdsRepositoryTests.swift
@@ -8,30 +8,8 @@
 
 @testable import desafio
 import XCTest
-import NetworkLayer
 
 class APIAdsRepositoryTests: XCTestCase {
-    
-    // MARK: Spy
-    
-    class NetworkProviderSpy: NetworkService {
-        var requestResult: Result<ListAdsDTO, NetworkError> = .failure(.noJSONData)
-        
-        func performRequest<T: Decodable>(
-            endpoint: Endpoint,
-            using keyDecodingStrategy: JSONDecoder.KeyDecodingStrategy = .convertFromSnakeCase,
-            result: @escaping ResultHandler<T>
-        ) {
-            
-            switch requestResult {
-            case let .success(ads):
-                //swiftlint:disable:next force_cast
-                result(.success(ads as! T))
-            case let .failure(error):
-                result(.failure(error))
-            }
-        }
-    }
     
     // MARK: Subject under test
